back/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["arseniy <a.moguchii@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.6.1"
fastapi = {extras = ["all"], version = "^0.115.5"}
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
loguru = "^0.7.2"
asyncpg = "^0.30.0"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-jose = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
